requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2